+++ CHANGELOG.md
@@ -602,3 +602,4 @@
 * Minor cleanup to GetEffectiveTheme and PanelController
 * EditorConditional and EditorMicroscript now throws if type is unknown
 * Player prefab name can now be overridden from WorldSceneController or GameState
+* Added PersistInitialPosition option to ActorController to save InitialPosition